+++ tests/features/test_smarterdb.py
@@ -116,8 +116,11 @@ class BreedTestCase(MongoMockMixin, unittest.TestCase):
 class VariantMixin():
     @classmethod
     def setUpClass(cls):
-        with open(DATA_DIR / "variants.json") as handle:
+        with open(DATA_DIR / "affy_variants.json") as handle:
             cls.data = json.load(handle)[0]
+
+        with open(DATA_DIR / "affy_variants.json") as handle:
+            cls.affy_data = json.load(handle)[2]
 
         super().setUpClass()
 
@@ -126,6 +129,9 @@ class LocationTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
     def setUp(self):
         location = self.data["locations"][1]
         self.location = Location.from_json(json.dumps(location))
+
+        affy_location = self.affy_data["locations"][2]
+        self.affy_location = Location.from_json(json.dumps(affy_location))
 
     def test_illumina_top(self):
         self.assertEqual(self.location.illumina_top, "A/G")
@@ -340,6 +346,44 @@ class LocationTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
             ["A", "T"]
         )
 
+    def test_is_affy(self):
+        for genotype in ["T/C", "T/T", "C/T", "C/C", "0/0"]:
+            genotype = genotype.split("/")
+
+            self.assertTrue(
+                self.affy_location.is_affymetrix(genotype),
+                msg=f"{genotype} is not in affymetrix coordinates!"
+            )
+
+        # is not in not if contains an allele not in top format
+        for genotype in ["A/A", "A/G", "G/A", "G/G"]:
+            self.assertFalse(
+                self.location.is_forward(genotype),
+                msg=f"{genotype} is in affymetrix coordinates!"
+            )
+
+    def test_affy2top(self):
+        """Test affymetrix to top conversion"""
+
+        affymetrixs = ["T/C", "T/T", "C/T", "C/C", "0/0"]
+        tops = ["A/G", "A/A", "G/A", "G/G", "0/0"]
+
+        for i, genotype in enumerate(affymetrixs):
+            reference = tops[i].split("/")
+            genotype = genotype.split("/")
+
+            test = self.affy_location.affy2top(genotype)
+            self.assertEqual(reference, test)
+
+    def test_affy2top_error(self):
+        """Test exception with an allele not in affymetrix coding"""
+        self.assertRaisesRegex(
+            SmarterDBException,
+            "is not in affymetrix coding",
+            self.affy_location.affy2top,
+            ["A", "T"]
+        )
+
 
 class VariantSheepTestCase(VariantMixin, MongoMockMixin, unittest.TestCase):
     def setUp(self):
@@ -454,37 +498,51 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
         self.smarter_id = "ITOA-TEX-000000001"
         self.original_id = "TEST"
 
-        self.sample = SampleSheep(
-            original_id=self.original_id,
-            smarter_id=None
-        )
-
-        # need country, breed and species in order to get a smarter_id
-        self.sample.country = "Italy"
-        self.sample.breed = "Texel"
-        self.sample.species = "Sheep"
-
         # fetch some values from database
         self.dataset = Dataset.objects.get(file="test.zip")
         self.breed = Breed.objects.get(species="Sheep", code="TEX")
+
+        # this is required to create a sample
+        self.country = "Italy"
+
+        # additional sample fields
+        self.chip_name = "IlluminaOvineSNP50"
+        self.sex = SEX.MALE
+        self.alias = "TEST-ALIAS"
 
     def tearDown(self):
         SampleSheep.objects().delete()
 
         super().tearDown()
 
-    def test__str(self):
+    def create_sample(self):
+        """Create a sample instance in mongodb"""
+
+        self.sample = SampleSheep(
+            original_id=self.original_id,
+            smarter_id=None
+        )
+
+        # need country, breed and species in order to get a smarter_id
+        self.sample.country = self.country
+        self.sample.breed = "Texel"
+        self.sample.species = "Sheep"
+
         # save sample in db
         self.sample.save()
+
+        return self.sample
+
+    def test__str(self):
+        self.create_sample()
 
         self.assertEqual(
             str(self.sample),
             f"{self.smarter_id} (Texel)"
         )
 
-    def test_save(self):
-        # save sample in db
-        self.sample.save()
+    def test_save_smarter_id(self):
+        self.create_sample()
 
         self.assertEqual(self.sample.smarter_id, self.smarter_id)
 
@@ -495,7 +553,11 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
             self.original_id,
             self.dataset,
             self.breed,
-            "Italy")
+            self.country,
+            self.chip_name,
+            self.sex,
+            self.alias
+        )
 
         self.assertIsInstance(sample, SampleSheep)
         self.assertEqual(sample.smarter_id, self.smarter_id)
@@ -508,7 +570,11 @@ class SampleSheepTestCase(SmarterIDMixin, MongoMockMixin, unittest.TestCase):
             self.original_id,
             self.dataset,
             self.breed,
-            "Italy")
+            self.country,
+            self.chip_name,
+            self.sex,
+            self.alias
+        )
 
         self.assertIsInstance(sample, SampleSheep)
         self.assertEqual(sample.smarter_id, self.smarter_id)